+++ src/error.rs
@@ -1,6 +1,6 @@
 use thiserror::Error;
 
-use crate::parser::{Error as ParseError, LSpan};
+use crate::parser::Error as ParseError;
 
 /// Error with a location info.
 pub trait LocatedError {
@@ -18,10 +18,6 @@ pub enum Error<'a, GE> {
     #[error("parse failed: {0}")]
     Parse(ParseError<'a>),
 
-    /// Invalid or unsupported tag.
-    #[error("unsupported tag: {0}")]
-    Tag(LSpan<'a>),
-
     /// Generating step failed, see document of generator type for detail.
     #[error("generator failed: {0}")]
     Gen(GE),
@@ -31,7 +27,6 @@ impl<'a, GE: LocatedError> LocatedError for Error<'a, GE> {
     fn location(&self) -> (usize, usize) {
         match self {
             Self::Parse(e) => e.location(),
-            Self::Tag(span) => (span.extra, span.get_column()),
             Self::Gen(e) => e.location(),
         }
     }