+++ src/generator/helper/mod.rs
@@ -1,10 +1,11 @@
 //! Helper functions for create generator.
 
 mod error;
+mod flatten;
 mod tag;
 mod unescape;
 
-
 pub use error::GeneratorInfallible;
+pub use flatten::flatten;
 pub use tag::{CustomTagParser, NoopCustomTagParser};
 pub use unescape::{unescape, Unescape};